Add support for DDS Union types
Prior to this commit there was no support for parsing DDS Unions.
As a user looking to parse OMG IDL files with union types I was
unable harness the power of the rest of the tool until unions
were supported. This commit represents the addition of parsing support
for union types defined in an OMG IDL file.

Because my use case benefits from the functionality of the
`refine_typename` method of the IDLNode class, I revised
some of its logic and the logic it requires then uncommented
its use in each of the modules that had disabled it.

Further, tests were added to the module_test.py suite to validate
the parser's handling of union types.