Add step to prepare input before parsing idl file
Prior to this commit this tool's parsing logic made assumptions about
whitespace inside array and sequence definitions. Due to these
assumptions the following scenarios broke the tool's ability
to correctly parse a legal OMG IDL file:

```
int array1[ 10 ];
int array2[
  10
  ];
sequence< int > seq1;
sequence<
  int
  > seq2;
```

Logic was added to prepare a file to meet the assumptions of
the parser before parsing. With regular expressions any whitespace
between brackets (including newlines) is eliminated.

Another assumption made by the parser was that arrays were defined
with integer constants. This resulted in the following scenario
breaking the tool's parsing ability:

```
const MY_CONSTANT = 10;

int array1[ MY_CONSTANT ];

```

After this commit the tool no longer makes the assumption that the
size of an array is defined by and integer constant. This resulted
in a modification of the code that cast the value to an `int` as
well as modifications to tests that asserted the value was an int
value.

I was also able to find certain scenarios that challenged the
assumptions I made about the refine_typename function in a
previous commit. Small changes to that function resolved the
issues I encountered.