Updated array size handling logic.
Prior to this commit the this branch treated the defined
size of an array as a string. This was problematic as it
forced the user to parse the string to get the literal value
associated with the array size.

This commit adds logic to check whether the size of the array
is defined with a constant value and sets the array value to
an integer value rather than a string.

A little extra validation was added to ensure the specified
size is sensical (if it is not a constant it should be able to
be converted to an integer)

Tests were updated to assure the array sizes are defined as
integers and that the correct constant values are associated
with array size.

Before pushing these changes, I passed the file `basic_module_test.idl`
through the Vortex DDS compiler `idlpp` and some minor details
were addressed to ensure the file compiled sucessfully.