Account for IDL files without explicit modules.
Prior to this commit logic to retrieve the size of an array
defined in an IDL file assumed that one or more modules
we explicitly defined. This resulted in a crash when trying
to access the parser's non-existant list of modules (index out of range).

After this commit the parser does not assume that there are explicit
modules to reference.